Finalized navmesh code base
A ton of changes went into these to get it in working order, and I can now build a navmesh using Valve's generator and correctly load and use it with custom data.

The List class was changed from being an abstract object and now inherits from Valves Vector class as a base, so that I can have some functions that Valve doesn't have.

The loader was changed from being an abstract object as it didn't make sense to allocate memory to a pointer that will only be used once at run time, so it's now static functions, yes, plural, I added a Save function.

I had to fiddle with the connection list and ladder list that is stored within NavArea's so that I can save those properly, they are now a list of lists of objects.

Some classes got a Destroy() function so that I can de-allocate without using a destructor, specifically classes that contained lists of objects. The NavMesh class doesn't have a destroy as I intend to use its destructor since I will be calling delete on an INavMesh pointer when the map ends.

I added a new class to be used as a hint so that I can save some processing on calculating suitable locations to setup engineer nests and sniper camps. I also completed the grid class so that it doesn't use CellArrays and is instead a list of lists of areas.