Some fixes and optimizations
Hooked `IServerPluginHelpers::ClientCommand()` to rate limit it so bots don't get kicked for too many commands following a recent lag-bot-prevention patch to TF2, hooked this function because I use it everywhere

Added ConVar change callbacks to `bot_enabled` and `bot_debug` so things proactively update on change

Added some logic to handle ClientCommand on old engines, even though this probably won't be used on engines old enough for it to matter. Old versions didn't have the command arguments passed into the function and instead required calling into the engine to get at them.

Move GameRules creation to OnAllLoaded since that's when SDKTools actually creates it, use a frame action to wait a small moment for it to be valid

Add a game config offset for the control point data offset and a keyvalue for the GameRules net class for game specific lookup

Move the FireEvent from here to the events manager, handle everything internally