Optimizations
Pass around pointers to the lists instead of copy constructing.

Constify all Vector returns because we don't want to change them... ever.

Remove ForEachItem macro, intellisense doesn't like it and I added range looping support anyway which looks cleaner

Added a Verbose version of the debug function which follows a convar, since it's very slow to log every little bit of loading the navmesh but I want to have the option if I need to